Remove runtimeconfiguration
