+++ packages/sanes-chrome-extension/src/logic/config/index.ts
@@ -1,3 +1,2 @@
 export * from './codec';
 export * from './configurationfile';
-export * from './runtimeconfiguration';
